update: allow update absent nullable fields
Update operations can't insert with gaps.
But there exist "virtual nulls" in tarantool -
i.e. fields that are specified in format and marked as is_nullable.
They are not actually stored in the database, and currently
tarantool refuses to do update ops with such nulls as "gaps".

But from the user side these nulls exist. So this patch changes
behaviour to such that update ops will fill virtual nulls with
the real ones.

Closes #3378