box/func: fix modules functions restore
In commit 96938fafb an ability to hot reload of modules
has been introduced. When module is been reloaded his
functions are resolved to new symbols but if something
went wrong it is supposed to restore old symbols from
the old module. Actually current code restores only
one function and may crash if there a bunch of functions
to restore. Lets fix it.

Part-of #4642

Signed-off-by: Cyrill Gorcunov <gorcunov@gmail.com>